fix(admin): chỉnh header layout và style

+++ src/layouts/admin/Header.jsx
@@ -14,7 +14,9 @@ import {
   Menu,
 } from "lucide-react";
 import "../../styles/admin/header.css";
-
+import { useLoading } from "../../hooks/useLoading.js";
+import { useAuth } from "../../hooks/useAuth.js";
+import { handleApiError, handleApiSuccess } from "../../utils/apiHandler.js";
 const ROUTE_TITLES = [
   ["/admin/product-groups", "Nhóm sản phẩm"],
   ["/admin/products/colors", "Sản phẩm / Màu sắc"],
@@ -43,23 +45,27 @@ export default function Header({ toggleSidebar }) {
   const parent = parts.length > 1 ? parts[0] : "Trang chủ";
   const current = parts[parts.length - 1];
 
+  const { callAuth } = useAuth();
+  const { loading } = useLoading();
+
   const handleLogout = async () => {
+    if (loading) return;
+
     try {
-      const res = await logoutAPI();
+      await callAuth(() => logoutAPI(), {
+        showLoading: false,
+      });
+
       localStorage.clear();
       setUser(null);
-      api.success({
-        message: "Đăng xuất thành công",
-        description: res?.message,
-      });
-      setTimeout(() => navigate("/"), 500);
-    } catch (e) {
-      api.error({
-        message: "Lỗi",
-        description: e?.message,
-      });
-    } finally {
+
+      handleApiSuccess(api, "Đăng xuất thành công!");
+
       setDropOpen(false);
+
+      navigate("/");
+    } catch (err) {
+      handleApiError(api, err);
     }
   };
